fixing env and db prover insertion

+++ src/database/prover.rs
@@ -1,10 +1,7 @@
-use sqlx::FromRow;
-
 pub type Provers = Vec<Prover>;
 
-#[derive(FromRow)]
 pub struct Prover {
     pub url:        String,
-    pub batch_size: i64,
-    pub timeout_s:  i64,
+    pub batch_size: usize,
+    pub timeout_s:  u64,
 }
